hot fix

+++ lib/model/History.dart
@@ -30,7 +30,7 @@ class His {
   String payDate;
   String payTime;
   int payNum;
-  int paySale;
+  double paySale;
   String payBank;
   int orId;
   String orDate;
